types

+++ code/model_lbd.jl
@@ -11,7 +11,7 @@ include("postprocess.jl")
 """Initialize prices"""
 function init_PV(game)::Tuple{Matrix{Float64}, Matrix{Float64}}
     if game.policy == "limitedmergers"
-        filename = string("nobundling_a", Int64(floor(game.alpha*100)), "g", Int64(floor(game.gamma*100)), "s", Int64(floor(game.sigma*10)), ".json")
+        filename = string("nobundling_a", Int(floor(game.alpha*100)), "g", Int(floor(game.gamma*100)), "s", Int(floor(game.sigma*10)), ".json")
         temp = init.import_game(filename)
         P = temp.P
         V = game.V
@@ -44,7 +44,7 @@ function correct_P(game, P::Matrix{Float64})::Matrix{Float64}
 end
 
 """Precompute stuff"""
-function precompute(game, row::Int64, W::Array{Float64,3})
+function precompute(game, row::Int, W::Array{Float64,3})
 
     # Check active firms
     active_n = game.active_firms[row, :]
@@ -65,7 +65,7 @@ function precompute(game, row::Int64, W::Array{Float64,3})
     w_signed = copy(w)
     w_signed += W[row, active_out, game.partner[active_n[1:4]]]' .* (partner.>0)
     w_signed .*= sign
-    joint_p = Int8.([1,2,1,2][active_n[1:4]])
+    joint_p = Int.([1,2,1,2][active_n[1:4]])
 
     return active_n, out, mc, outcomes, partner, w, w_signed, joint_p
 end
@@ -81,7 +81,7 @@ function demand(p::Vector{Float64}, sigma::Float64, p0::Float64, outcomes, out):
 end
 
 """Compute profits for BR iteration"""
-function BR(x::Vector{Float64}, n::Int64, p::Vector{Float64}, sigma::Float64, p0::Float64, mc::Vector{Float64}, outcomes, partner::Vector{Int8}, w, out)::Float64
+function BR(x::Vector{Float64}, n::Int, p::Vector{Float64}, sigma::Float64, p0::Float64, mc::Vector{Float64}, outcomes, partner::Vector{Int}, w, out)::Float64
     p[n] = x[1]                                         # Insert price of firm n
     q, d = demand(p, sigma, p0, outcomes, out)# Compute demand
     V = (p .- mc) .* d + sum(w .* q, dims=2)            # Compute value
@@ -90,7 +90,7 @@ function BR(x::Vector{Float64}, n::Int64, p::Vector{Float64}, sigma::Float64, p0
 end
 
 """Compute price by best reply itearation"""
-function update_p_BR(game, p::Vector{Float64}, row::Int64, W::Array{Float64,3})::Vector{Float64}
+function update_p_BR(game, p::Vector{Float64}, row::Int, W::Array{Float64,3})::Vector{Float64}
 
     # Precompute things
     active_n, out, mc, outcomes, partner, w, w_signed, joint_p = precompute(game, row, W)
@@ -122,7 +122,7 @@ function update_P_BR(game, W::Array{Float64,3})::Matrix{Float64}
 end
 
 """First order condition"""
-function FOC(p::Vector{Float64}, sigma::Float64, p0::Float64, mc::Vector{Float64}, outcomes, partner::Vector{Int8}, w_signed::Matrix{Float64}, joint_p::Vector{Int8}, out)::Vector{Float64}
+function FOC(p::Vector{Float64}, sigma::Float64, p0::Float64, mc::Vector{Float64}, outcomes, partner::Vector{Int}, w_signed::Matrix{Float64}, joint_p::Vector{Int}, out)::Vector{Float64}
 
     # Compute demand
     q, d = demand(p, sigma, p0, outcomes, out)
@@ -145,7 +145,7 @@ function FOC(p::Vector{Float64}, sigma::Float64, p0::Float64, mc::Vector{Float64
 end
 
 """Update p by numerically solving first order condition"""
-function update_p_FOC(game, W::Array{Float64,3}, row::Int64)::Vector{Float64}
+function update_p_FOC(game, W::Array{Float64,3}, row::Int)::Vector{Float64}
 
     # Precompute things
     active_n, out, mc, outcomes, partner, w, w_signed, joint_p = precompute(game, row, W)
